% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask}
\usage{
ask(id, content, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id of the input this defines.
The input that is defined returns \code{TRUE}
if the user accepts, and \code{FALSE} if
the user rejects. See details.}

\item{content}{Content of the prompt,
question to prompt.}

\item{...}{Options.}

\item{session}{A valid shiny session.}
}
\description{
Create a prompt: 'confirm' or 'cancel'.
}
\details{
The \code{id} is used to define the input
so \code{ask(id = "xxx")} will return the response
of at \code{input$xxx}.
}
