% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awn.R
\name{awn_globals}
\alias{awn_globals}
\title{Global Options}
\usage{
awn_globals(..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Options.}

\item{session}{A valid shiny session.}
}
\description{
Define global options.
}
