% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Notify}
\usage{
notify(
  content,
  type = c("tip", "info", "warning", "success", "alert"),
  ...,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{content}{Content to display.}

\item{type}{Type of notification.}

\item{...}{Options to pass to the notification.}

\item{session}{A valid shiny \code{session}.}
}
\description{
Show a notification.
}
